+++ packages/skin-bleu/src/filter.tsx
@@ -2,6 +2,7 @@ import { Box, Input } from "@mui/material";
 import { shallowEqual, useColProps, useColRef } from "@rttui/core";
 import React from "react";
 import { Resizer } from "./resizer";
+import { useSelectionManagerCls } from "./selection_manager_context";
 
 export function Filter({ resizer }: { resizer?: boolean }) {
   const { filterValue, canFilter } = useColProps({
@@ -15,6 +16,7 @@ export function Filter({ resizer }: { resizer?: boolean }) {
     areCallbackOutputEqual: shallowEqual,
   });
   const colRef = useColRef();
+  const selectionManager = useSelectionManagerCls();
 
   if (!canFilter && !resizer) {
     return null;
@@ -36,7 +38,11 @@ export function Filter({ resizer }: { resizer?: boolean }) {
             minWidth: "120px",
           }}
           value={filterValue}
+          onFocus={() => {
+            selectionManager.blur();
+          }}
           onChange={(e) => {
+            e.preventDefault();
             React.startTransition(() => {
               colRef().column?.setFilterValue(e.target.value);
             });