+++ packages/skin-tailwind/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rttui/skin-tailwind
 
+## 0.0.49
+
+### Patch Changes
+
+- Updated dependencies
+  - @rttui/core@1.0.54
+
 ## 0.0.48
 
 ### Patch Changes